# openapi_spec.yaml
openapi: 3.0.3
info:
  title: Docaiche API
  description: |
    AI-powered documentation search and content management system with advanced analytics capabilities.
    
    ## Overview
    Docaiche provides intelligent documentation search, content enrichment, and comprehensive analytics
    for tracking user behavior and system performance. The API supports real-time search with caching,
    content ingestion from multiple sources, and AI-powered content enhancement.
    
    ## Key Features
    - **Intelligent Search**: AI-powered search with technology filtering and relevance scoring
    - **Content Management**: Ingest and manage documentation from various sources
    - **Analytics Dashboard**: Comprehensive metrics for search patterns, user behavior, and system performance
    - **AI Enrichment**: Automatic content enhancement using LLM providers
    - **Real-time Monitoring**: Live system health and performance metrics
    
    ## Analytics Capabilities
    The analytics endpoints provide rich data for building beautiful dashboards:
    - Search query trends and patterns over time
    - User engagement metrics (clicks, dwell time, scroll depth)
    - Content quality scores and performance
    - System performance metrics (response times, cache hit rates)
    - Technology adoption trends
    - Real-time activity feeds
    
    ## Monitoring Integration
    - **Loki**: Centralized log aggregation for all services
...
    
  version: 2.0.0
  contact:
...
  license:
    name: MIT
...
servers:
...
paths:
...
  /config:
    get:
      tags: [config]
      summary: Get configuration
      description: List all configuration items
      operationId: getConfiguration
      parameters:
        - name: category
          in: query
          description: Filter by category
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Configuration items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigurationResponse'
    
    put:
      tags: [config]
      summary: Update configuration
      description: Update a configuration value
      operationId: updateConfiguration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigUpdateRequest'
      responses:
        '202':
          description: Configuration update accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigUpdateResponse'
...
components:
  schemas:
...
    ConfigurationResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ConfigItem'
    ConfigItem:
      type: object
      required: [key, value, schema_version, is_sensitive]
      properties:
        key:
          type: string
          description: Configuration key
        value:
          description: Configuration value (type varies)
        schema_version:
          type: string
          description: Schema version
        description:
          type: string
          description: Configuration description
        is_sensitive:
          type: boolean
          description: Whether value is sensitive
    ConfigUpdateRequest:
      type: object
      required: [key, value]
      properties:
        key:
          type: string
          description: Configuration key to update
        value:
          description: New configuration value
    ConfigUpdateResponse:
      type: object
      required: [message, key, status]
      properties:
        message:
          type: string
        key:
          type: string
        status:
          type: string
          enum: [accepted, rejected]
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (future implementation)
security: []  # No security currently implemented